Extract src/update.ts — app self-update and the version label
First slice of the bootstrap-trim item, chosen because it is the one cluster
left that needs *nothing* from main.ts: the only thing it asks about the rest
of the app is how many live panes an install would kill, and that is ./state.

So it takes no hook — and exports nothing either. `import "./update"` for its
side effects is the whole of the wiring: it owns its footer chip, its two
click handlers, the quiet check 3s after launch and the tray-check-updates
listener. That is a new pattern in this repo, so the import says so; an import
with no bindings otherwise reads like something an extraction forgot to clean
up.

Verified mechanical in the strongest sense available: the relocated block
diffed against `git show HEAD:src/main.ts` is byte-identical — not one export
keyword, not one host prefix.

PLAN now records the order for what remains, by independence rather than file
order: the external+dormant block (which is where the read-only render mirrors
the last item left behind belong), then the run-on-stop/task-run block (which
will *remove* hooks, since inspector.ts and tasks.ts currently reach into it),
then the pane layer last.

main.ts 1,593 -> 1,514. Gates: tsc --noEmit clean, 368 vitest pass. No Rust
touched.
